Programs/Java/HW08 Residence Listings/src/Residence.java
/**
 * Created by Todd on 10/31/2016.
 */
public class Residence {

    private String address;
    private int numBedrooms;
    private int numBathrooms;
    private int squareFootage;
    private double monthlyRent;

    public Residence (String address, int numBedrooms, int numBathrooms, int squareFootage, double monthlyRent) {
        this.address = address;
        this.numBedrooms = numBedrooms;
        this.numBathrooms = numBathrooms;
        this.squareFootage = squareFootage;
        this.monthlyRent = monthlyRent;
    }

    public String getAddress () {
        return address;
    }

    public int getNumBedrooms () {
        return numBedrooms;
    }

    public int getNumBathrooms () {
        return numBathrooms;
    }

    public int getSquareFootage () {
        return squareFootage;
    }

    public double getMonthlyRent () {
        return monthlyRent;
    }
}
